refactor(project-files): extract query owner
