+++ scripts/ci/module-impact.json
@@ -119,6 +119,7 @@
       "ownerPaths": [
         "src/main/project-files/mutation-owner.ts",
         "src/main/project-files/mutation-projection.ts",
+        "src/main/project-files/query-owner.ts",
         "src/main/project-files/query-support.ts",
         "src/main/project-files/repository.ts"
       ],
