+++ src/main/project-files/repository.ts
@@ -1,12 +1,9 @@
-import { Prisma, type ManagedFile } from '@prisma/client'
-
 import type { PersistedChatSession } from '../../shared/session-persistence'
 import type {
   ArtifactGroupPage,
   GetProjectFilesOverviewRequest,
   ListArtifactGroupsRequest,
   ListProjectFilesRequest,
-  ProjectFileItem,
   ProjectFilesOverview,
   ProjectFilesPage,
   ProjectFileSource,
@@ -23,35 +20,19 @@ import type {
   ProjectFilesClientFactory,
   ProjectFilesClientProvider
 } from './mutation-projection'
-import {
-  countMatchingArtifacts,
-  decodeFileCursor,
-  decodeGroupCursor,
-  decodeSearchArtifactCursor,
-  encodeCursor,
-  getMatchingOverviewCounts,
-  listMatchingArtifactGroups,
-  listMatchingArtifacts,
-  listMatchingFiles,
-  listOtherProjectArtifacts,
-  normalizeLimit,
-  normalizeSearch,
-  requireIdentifier,
-  toOriginProjection,
-  toProjectFileItem,
-  toSafeCount
-} from './query-support'
+import { ProjectFilesQueryOwner } from './query-owner'
 
 // Stable public facade for the query-optimized Project Files projection. Session JSON remains
-// authoritative; one internal owner holds every mutation and completeness state transition.
+// authoritative; internal owners separate read orchestration from mutation/completeness state.
 class ManagedFileIndexRepository {
   private readonly mutationOwner: ProjectFilesMutationOwner
+  private readonly queryOwner: ProjectFilesQueryOwner
 
-  constructor(
-    private readonly getClient: ProjectFilesClientProvider,
-    private readonly dataRoot: string
-  ) {
+  constructor(getClient: ProjectFilesClientProvider, dataRoot: string) {
     this.mutationOwner = new ProjectFilesMutationOwner(getClient, dataRoot)
+    this.queryOwner = new ProjectFilesQueryOwner(getClient, dataRoot, (projectId) =>
+      this.mutationOwner.isIndexComplete(projectId)
+    )
   }
 
   syncSession(
@@ -92,288 +73,19 @@ class ManagedFileIndexRepository {
   async getOverview(
     request: string | GetProjectFilesOverviewRequest
   ): Promise<ProjectFilesOverview> {
-    const { projectId, search: rawSearch } =
-      typeof request === 'string' ? { projectId: request, search: undefined } : request
-    requireIdentifier(projectId, 'projectId')
-    const search = normalizeSearch(rawSearch)
-    const client = await this.getClient()
-    const [totalCount, uploadCount, artifactCount, artifactGroupCount] = search
-      ? await getMatchingOverviewCounts(client, projectId, search)
-      : await Promise.all([
-          client.managedFile.count({ where: { projectId, deletedAt: null } }),
-          client.managedFile.count({ where: { projectId, source: 'upload', deletedAt: null } }),
-          client.managedFile.count({ where: { projectId, source: 'artifact', deletedAt: null } }),
-          client.managedFileSessionSync.count({
-            where: { projectId, deletedAt: null, artifactCount: { gt: 0 } }
-          })
-        ])
-
-    return {
-      totalCount,
-      uploadCount,
-      artifactCount,
-      artifactGroupCount,
-      isIndexComplete: this.mutationOwner.isIndexComplete(projectId)
-    }
+    return this.queryOwner.getOverview(request)
   }
 
   async listFiles(request: ListProjectFilesRequest): Promise<ProjectFilesPage> {
-    requireIdentifier(request.projectId, 'projectId')
-    const collection = request.collection as { kind?: unknown; sessionId?: unknown }
-    let normalizedCollection: ListProjectFilesRequest['collection']
-    if (collection.kind === 'all') {
-      normalizedCollection = { kind: 'all' }
-    } else if (collection.kind === 'uploads') {
-      normalizedCollection = { kind: 'uploads' }
-    } else if (collection.kind === 'sessionArtifacts' && typeof collection.sessionId === 'string') {
-      requireIdentifier(collection.sessionId, 'sessionId')
-      normalizedCollection = { kind: 'sessionArtifacts', sessionId: collection.sessionId }
-    } else {
-      throw new Error('Project files collection is invalid.')
-    }
-    const normalizedRequest = { ...request, collection: normalizedCollection }
-    const client = await this.getClient()
-    const limit = normalizeLimit(request.limit)
-    const search = normalizeSearch(request.search)
-    const source =
-      normalizedCollection.kind === 'all'
-        ? undefined
-        : normalizedCollection.kind === 'uploads'
-          ? 'upload'
-          : 'artifact'
-    const sessionId =
-      normalizedCollection.kind === 'sessionArtifacts' ? normalizedCollection.sessionId : undefined
-    if (sessionId && search?.excludedSessionIds.includes(sessionId)) {
-      return { items: [], totalCount: 0 }
-    }
-    const cursor = request.cursor ? decodeFileCursor(request.cursor, normalizedRequest) : undefined
-    const where: Prisma.ManagedFileWhereInput = {
-      projectId: request.projectId,
-      ...(source ? { source } : {}),
-      deletedAt: null,
-      ...(sessionId !== undefined
-        ? { sessionId }
-        : search?.excludedSessionIds.length
-          ? { sessionId: { notIn: search.excludedSessionIds } }
-          : {}),
-      ...(cursor
-        ? {
-            OR: [
-              { sortAtMs: { lt: BigInt(cursor.sortAtMs) } },
-              { sortAtMs: BigInt(cursor.sortAtMs), seq: { lt: cursor.seq } }
-            ]
-          }
-        : {})
-    }
-    const [rows, totalCount] = search
-      ? await listMatchingFiles(client, request.projectId, source, sessionId, search, cursor, limit)
-      : await Promise.all([
-          client.managedFile.findMany({
-            where,
-            orderBy: [{ sortAtMs: 'desc' }, { seq: 'desc' }],
-            take: limit + 1
-          }),
-          client.managedFile.count({
-            where: {
-              projectId: request.projectId,
-              ...(source ? { source } : {}),
-              deletedAt: null,
-              ...(sessionId !== undefined ? { sessionId } : {})
-            }
-          })
-        ])
-    const pageRows = rows.slice(0, limit)
-    const lastRow = pageRows.at(-1)
-    const origins = await client.fileOriginSession.findMany({
-      where: {
-        projectId: request.projectId,
-        sessionId: { in: [...new Set(pageRows.map((row) => row.sessionId))] }
-      }
-    })
-    const originsBySession = new Map(origins.map((origin) => [origin.sessionId, origin]))
-
-    return {
-      items: pageRows.map((row) =>
-        toProjectFileItem(row, this.dataRoot, originsBySession.get(row.sessionId))
-      ),
-      totalCount,
-      nextCursor:
-        rows.length > limit && lastRow
-          ? encodeCursor({
-              version: 2,
-              kind: normalizedCollection.kind,
-              projectId: request.projectId,
-              sessionId,
-              queryKey: search?.queryKey ?? '',
-              sortAtMs: lastRow.sortAtMs.toString(),
-              seq: lastRow.seq
-            })
-          : undefined
-    }
+    return this.queryOwner.listFiles(request)
   }
 
   async searchArtifacts(request: SearchArtifactsRequest): Promise<SearchArtifactsResult> {
-    requireIdentifier(request.primaryProjectId, 'primaryProjectId')
-    if (!Array.isArray(request.otherProjectIds)) {
-      throw new Error('Project files otherProjectIds must be an array.')
-    }
-    const otherProjectIds = [...new Set(request.otherProjectIds)]
-      .filter((projectId) => projectId !== request.primaryProjectId)
-      .map((projectId) => {
-        requireIdentifier(projectId, 'otherProjectId')
-        return projectId
-      })
-    if (!Number.isInteger(request.otherLimit) || request.otherLimit < 0 || request.otherLimit > 5) {
-      throw new Error('Project files otherLimit must be between 0 and 5.')
-    }
-
-    const primaryLimit = normalizeLimit(request.primaryLimit)
-    const search = normalizeSearch({
-      filenameContains: request.filenameContains ?? '',
-      ...(request.excludedSessionIds === undefined
-        ? {}
-        : { excludedSessionIds: request.excludedSessionIds })
-    })
-    const cursor = request.primaryCursor
-      ? decodeSearchArtifactCursor(request.primaryCursor, request.primaryProjectId, search)
-      : undefined
-    const client = await this.getClient()
-    const excludedSessionIds = search?.excludedSessionIds ?? []
-    const [primaryRows, primaryTotalCount, otherRows] = await Promise.all([
-      listMatchingArtifacts(
-        client,
-        request.primaryProjectId,
-        search,
-        excludedSessionIds,
-        cursor,
-        primaryLimit
-      ),
-      countMatchingArtifacts(client, request.primaryProjectId, search, excludedSessionIds),
-      request.otherLimit > 0 && otherProjectIds.length > 0
-        ? listOtherProjectArtifacts(
-            client,
-            otherProjectIds,
-            search,
-            excludedSessionIds,
-            request.otherLimit
-          )
-        : Promise.resolve([])
-    ])
-    const primaryPageRows = primaryRows.slice(0, primaryLimit)
-    const lastPrimaryRow = primaryPageRows.at(-1)
-    const rows = [...primaryPageRows, ...otherRows]
-    const origins =
-      rows.length === 0
-        ? []
-        : await client.fileOriginSession.findMany({
-            where: {
-              OR: [
-                ...new Map(rows.map((row) => [`${row.projectId}:${row.sessionId}`, row])).values()
-              ].map((row) => ({ projectId: row.projectId, sessionId: row.sessionId }))
-            }
-          })
-    const originsBySession = new Map(
-      origins.map((origin) => [`${origin.projectId}:${origin.sessionId}`, origin])
-    )
-    const toItem = (row: ManagedFile): ProjectFileItem =>
-      toProjectFileItem(
-        row,
-        this.dataRoot,
-        originsBySession.get(`${row.projectId}:${row.sessionId}`)
-      )
-
-    return {
-      primary: {
-        items: primaryPageRows.map(toItem),
-        totalCount: primaryTotalCount,
-        nextCursor:
-          primaryRows.length > primaryLimit && lastPrimaryRow
-            ? encodeCursor({
-                version: 2,
-                kind: 'globalArtifacts',
-                primaryProjectId: request.primaryProjectId,
-                queryKey: search?.queryKey ?? '',
-                sortAtMs: lastPrimaryRow.sortAtMs.toString(),
-                seq: lastPrimaryRow.seq
-              })
-            : undefined
-      },
-      other: otherRows.map(toItem),
-      isIndexComplete: [request.primaryProjectId, ...otherProjectIds].every((projectId) =>
-        this.mutationOwner.isIndexComplete(projectId)
-      )
-    }
+    return this.queryOwner.searchArtifacts(request)
   }
 
   async listArtifactGroups(request: ListArtifactGroupsRequest): Promise<ArtifactGroupPage> {
-    requireIdentifier(request.projectId, 'projectId')
-    const client = await this.getClient()
-    const limit = normalizeLimit(request.limit)
-    const search = normalizeSearch(request.search)
-    const cursor = request.cursor ? decodeGroupCursor(request.cursor, request) : undefined
-    const groupWhere: Prisma.ManagedFileSessionSyncWhereInput = {
-      projectId: request.projectId,
-      deletedAt: null,
-      artifactCount: { gt: 0 },
-      ...(search?.excludedSessionIds.length
-        ? { sessionId: { notIn: search.excludedSessionIds } }
-        : {})
-    }
-    const where: Prisma.ManagedFileSessionSyncWhereInput = {
-      ...groupWhere,
-      ...(cursor
-        ? {
-            OR: [
-              { groupSortAtMs: { lt: BigInt(cursor.groupSortAtMs) } },
-              {
-                groupSortAtMs: BigInt(cursor.groupSortAtMs),
-                sessionId: { lt: cursor.sessionId }
-              }
-            ]
-          }
-        : {})
-    }
-    const [rows, totalCount] = search
-      ? await listMatchingArtifactGroups(client, request.projectId, search, cursor, limit)
-      : await Promise.all([
-          client.managedFileSessionSync.findMany({
-            where,
-            orderBy: [{ groupSortAtMs: 'desc' }, { sessionId: 'desc' }],
-            take: limit + 1
-          }),
-          client.managedFileSessionSync.count({
-            where: groupWhere
-          })
-        ])
-    const pageRows = rows.slice(0, limit)
-    const lastRow = pageRows.at(-1)
-    const origins = await client.fileOriginSession.findMany({
-      where: {
-        projectId: request.projectId,
-        sessionId: { in: pageRows.map((row) => row.sessionId) }
-      }
-    })
-    const originsBySession = new Map(origins.map((origin) => [origin.sessionId, origin]))
-
-    return {
-      items: pageRows.map((row) => ({
-        sessionId: row.sessionId,
-        artifactCount: toSafeCount(row.artifactCount, 'artifact group count'),
-        ...toOriginProjection(originsBySession.get(row.sessionId))
-      })),
-      totalCount,
-      nextCursor:
-        rows.length > limit && lastRow
-          ? encodeCursor({
-              version: 2,
-              kind: 'artifactGroups',
-              projectId: request.projectId,
-              queryKey: search?.queryKey ?? '',
-              groupSortAtMs: lastRow.groupSortAtMs.toString(),
-              sessionId: lastRow.sessionId
-            })
-          : undefined
-    }
+    return this.queryOwner.listArtifactGroups(request)
   }
 }
